+++ testing/enterprise_users_mock.go
@@ -41,7 +41,7 @@ func (m *MockEnterpriseUsersServiceInterface) EXPECT() *MockEnterpriseUsersServi
 }
 
 // Disable2FAForEnterpriseUser mocks base method.
-func (m *MockEnterpriseUsersServiceInterface) Disable2FAForEnterpriseUser(gid any, uid int, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
+func (m *MockEnterpriseUsersServiceInterface) Disable2FAForEnterpriseUser(gid any, uid int64, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
 	m.ctrl.T.Helper()
 	varargs := []any{gid, uid}
 	for _, a := range options {
@@ -73,19 +73,19 @@ func (c *MockEnterpriseUsersServiceInterfaceDisable2FAForEnterpriseUserCall) Ret
 }
 
 // Do rewrite *gomock.Call.Do
-func (c *MockEnterpriseUsersServiceInterfaceDisable2FAForEnterpriseUserCall) Do(f func(any, int, ...gitlab.RequestOptionFunc) (*gitlab.Response, error)) *MockEnterpriseUsersServiceInterfaceDisable2FAForEnterpriseUserCall {
+func (c *MockEnterpriseUsersServiceInterfaceDisable2FAForEnterpriseUserCall) Do(f func(any, int64, ...gitlab.RequestOptionFunc) (*gitlab.Response, error)) *MockEnterpriseUsersServiceInterfaceDisable2FAForEnterpriseUserCall {
 	c.Call = c.Call.Do(f)
 	return c
 }
 
 // DoAndReturn rewrite *gomock.Call.DoAndReturn
-func (c *MockEnterpriseUsersServiceInterfaceDisable2FAForEnterpriseUserCall) DoAndReturn(f func(any, int, ...gitlab.RequestOptionFunc) (*gitlab.Response, error)) *MockEnterpriseUsersServiceInterfaceDisable2FAForEnterpriseUserCall {
+func (c *MockEnterpriseUsersServiceInterfaceDisable2FAForEnterpriseUserCall) DoAndReturn(f func(any, int64, ...gitlab.RequestOptionFunc) (*gitlab.Response, error)) *MockEnterpriseUsersServiceInterfaceDisable2FAForEnterpriseUserCall {
 	c.Call = c.Call.DoAndReturn(f)
 	return c
 }
 
 // GetEnterpriseUser mocks base method.
-func (m *MockEnterpriseUsersServiceInterface) GetEnterpriseUser(gid any, uid int, options ...gitlab.RequestOptionFunc) (*gitlab.User, *gitlab.Response, error) {
+func (m *MockEnterpriseUsersServiceInterface) GetEnterpriseUser(gid any, uid int64, options ...gitlab.RequestOptionFunc) (*gitlab.User, *gitlab.Response, error) {
 	m.ctrl.T.Helper()
 	varargs := []any{gid, uid}
 	for _, a := range options {
@@ -118,13 +118,13 @@ func (c *MockEnterpriseUsersServiceInterfaceGetEnterpriseUserCall) Return(arg0 *
 }
 
 // Do rewrite *gomock.Call.Do
-func (c *MockEnterpriseUsersServiceInterfaceGetEnterpriseUserCall) Do(f func(any, int, ...gitlab.RequestOptionFunc) (*gitlab.User, *gitlab.Response, error)) *MockEnterpriseUsersServiceInterfaceGetEnterpriseUserCall {
+func (c *MockEnterpriseUsersServiceInterfaceGetEnterpriseUserCall) Do(f func(any, int64, ...gitlab.RequestOptionFunc) (*gitlab.User, *gitlab.Response, error)) *MockEnterpriseUsersServiceInterfaceGetEnterpriseUserCall {
 	c.Call = c.Call.Do(f)
 	return c
 }
 
 // DoAndReturn rewrite *gomock.Call.DoAndReturn
-func (c *MockEnterpriseUsersServiceInterfaceGetEnterpriseUserCall) DoAndReturn(f func(any, int, ...gitlab.RequestOptionFunc) (*gitlab.User, *gitlab.Response, error)) *MockEnterpriseUsersServiceInterfaceGetEnterpriseUserCall {
+func (c *MockEnterpriseUsersServiceInterfaceGetEnterpriseUserCall) DoAndReturn(f func(any, int64, ...gitlab.RequestOptionFunc) (*gitlab.User, *gitlab.Response, error)) *MockEnterpriseUsersServiceInterfaceGetEnterpriseUserCall {
 	c.Call = c.Call.DoAndReturn(f)
 	return c
 }
